# backend/travel_plan/specification/api.yaml
openapi: 3.0.3
info:
  title: Travel Plan API
  description: 이 문서는 여행 계획 서비스의 API 명세서입니다.
  version: 1.0.0
servers:
  - url: 'http://localhost:8080'
    description: Local Server(개발용)
tags:
  - name: plan
    description: 여행 계획 관련 API
paths:
  /v1/plans:
    post:
      tags:
        - plan
      summary: 여행 생성
      description: 여행을 생성합니다.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PlanPostRequest'
      responses:
        '201':
          description: 여행 생성 성공
          headers:
            ETag:
              description: 생성된 ETag 값
              schema:
                type: string
                example: "1dsafadsf"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Plan'
        '400':
          description: 잘못된 요청
        '500':
          description: 서버 에러
    get:
      tags:
        - plan
      summary: 여행 Id 목록 조회
      description: 여행 Id 목록을 조회합니다. 자기 자신의 여행만 조회할 수 있습니다.
      parameters:
        - name: ownerId
          in: query
          description: 사용자 ID
          required: true
          schema:
            type: string
            example: 1
        - name: page
          in: query
          description: 페이지 번호
          required: false
          schema:
            type: integer
            default: 0
            minimum: 0
            example: 1
        - name: size
          in: query
          description: 페이지 크기
          required: false
          schema:
            type: integer
            maximum: 100
            minimum: 5
            default: 10
            example: 10
      responses:
        '200':
          description: 여행 조회 성공
          content:
            application/json:
              schema:
                type: array
                example: ["1dsafadsf", "fif1fdjiaj"]
                items: # 여행 id string 목록
                  type: string
        '400':
          description: 잘못된 요청
        '403':
          description: 접근 권한 없음
        '500':
          description: 서버 에러
  /v1/plans/{planId}:
    get:
      tags:
        - plan
      summary: 여행 전체 정보 조회
      description: 여행의 전체 정보를 한번에 조회합니다.
      parameters:
        - name: planId
          in: path
          description: 여행 ID
          required: true
          schema:
            type: string
            example: 1
        - name: If-None-Match
          in: header
          description: 이전에 조회한 ETag 값
          required: false
          schema:
            type: string
            example: "1dsafadsf"
      responses:
        '200':
          description: 여행 조회 성공
          headers:
            ETag:
              description: 조회한 ETag 값
              schema:
                type: string
                example: "1dsafadsf"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Plan'
        '304':
          description: IF-NONE-MATCH 와 ETag 값이 일치하는 경우
        '400':
          description: 잘못된 요청
        '500':
          description: 서버 에러
    delete:
      tags:
        - plan
      summary: 여행 삭제
      description: 여행을 삭제합니다.
      parameters:
        - name: planId
          in: path
          description: 여행 ID
          required: true
          schema:
            type: string
            example: 1
      responses:
        '200':
          description: 여행 삭제 성공
        '400':
          description: 잘못된 요청
        '500':
          description: 서버 에러
    put:
      tags:
        - plan
      summary: 여행 수정
      description: 여행을 수정합니다. 컴포넌트가 새롭게 추가되는 경우, id값을 지정하지 않아야하고, id값이 지정되어있는 경우, 해당 id값을 가진 컴포넌트가 수정됩니다. 기존의 컴포넌트를 생략한 경우, 기존의 컴포넌트는 삭제됩니다.
      parameters:
        - name: planId
          in: path
          description: 여행 ID
          required: true
          schema:
            type: string
            example: 1
        - name: If-Match
          in: header
          description: 이전에 조회한 ETag 값
          required: false
          schema:
            type: string
            example: "1dsafadsf"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PlanPutRequest'
      responses:
        '200':
          description: 여행 수정 성공
          headers:
            ETag:
              description: 수정된 ETag 값
              schema:
                type: string
                example: "1dsafadsf"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Plan'
        '400':
          description: 잘못된 요청
        '412':
          description: Plan 내의 version 값이 일치하지 않는 경우
        '500':
          description: 서버 에러
components:
  schemas:
    PlanPostRequest:
      type: object
      properties:
        title:
          type: string
          description: 여행 제목
          required: true
          minLength: 1
          maxLength: 100
          example: 여행 제목
        dayCount:
          type: integer
          required: true
          description: 여행 일수
          minimum: 1
          maximum: 30
          example: 3
        startDate:
          type: string
          format: yyyy-MM-dd
          required: false
          description: 여행 시작 날짜, 오늘이거나 미래
          example: 2021-01-01
        placeIds:
          required: false
          type: array
          description: 장소 ID 목록, 최적 위치로 추가되어서 생성됩니다.
          items:
            type: string
            example: "1dsafadsf"
    PlanPutRequest:
      type: object
      properties:
        title:
          type: string
          description: 여행 제목
          example: 여행 제목
        startDate:
          type: string
          format: yyyy-MM-dd
          description: 여행 시작 날짜
          example: 2021-01-01
        route:
          $ref: '#/components/schemas/Route'
    Plan:
      type: object
      properties:
        planId:
          type: string
          description: 여행 ID
          example: 1
        title:
          type: string
          description: 여행 제목
          example: 여행 제목
        startDate:
          type: string
          format: yyyy-MM-dd
          description: 여행 시작 날짜
          example: 2021-01-01
        createdAt:
          type: string
          format: yyyy-MM-dd'T'HH:mm:ss.SSS'Z'
          description: 여행 생성 시간
          example: 2021-01-01T12:00:00.000Z
        updatedAt:
          type: string
          format: yyyy-MM-dd'T'HH:mm:ss.SSS'Z'
          description: 여행 변경 시간
          example: 2021-01-01T12:00:00.000Z
        version:
          type: string
          description: 수정 버전 식별자
          example: adflasjdflkajsdf
        route:
          $ref: '#/components/schemas/Route'
    Route:
      type: array
      description: 여행 경로
      items:
        oneOf:
          - $ref: '#/components/schemas/Day'
          - $ref: '#/components/schemas/WayPoint'
    Day:
      type: object
      properties:
        dayId:
          type: string
          description: 여행 일자 ID, 없으면 새로 생성하라는 의미
          example: iafi3j2
        wayPoints:
          type: array
          items:
            $ref: '#/components/schemas/WayPoint'
        _type:
          type: string
          description: 컴포넌트 타입
          example: DAY
    WayPoint:
      type: object
      properties:
        waypointId:
          type: string
          description: 경유지 ID, 없으면 새로 생성하라는 의미
          example: a1b2c3d4
        placeId:
          type: string
          description: 장소 ID
          example: a1b2c3d4
        memo:
          type: string
          description: 메모
          example: 메모
        time:
          type: string
          format: HH:mm
          description: 시간
          example: 12:00
        _type:
          type: string
          description: 컴포넌트 타입
          example: WAYPOINT